Tests/LibWeb/Text/input/css/asf-allowed-in-custom-function-descriptor.html

<!doctype html>
<style>
    @function --foo(--bar) {
        --baz: var(--bar);
        result: var(--baz);
    }
</style>
<div id="target"></div>
<script src="../include.js"></script>
<script>
    test(() => {
        println(document.styleSheets[0].cssRules[0].cssText);
    });
</script>
