Tests/LibWeb/Text/expected/css/asf-allowed-in-custom-function-descriptor.txt

@function --foo(--bar) { --baz: var(--bar); result: var(--baz); }
